+++ src/models/mod.rs
@@ -2,3 +2,4 @@ pub mod variant;
 pub mod build;
 pub mod cytoband;
 pub mod sample;
+pub mod consequence;